IT'S FINALLY TURNING IN A STABLE MANNER, STARTED FROM THE BOTTOM NOW WE DIGGING!!! Few kinks to still be worked out of course, namely that need to adjust the mutliplier to have it rotate perfectly and sometimes it detects light too early, need to somehow figure out how it's doing that.

+++ include/robot/control.h
@@ -14,7 +14,7 @@ void startMotorAndEncoderTest();
 void startDriveTest();
 void createDriveUntilNewTile(bool* onFirstTile);
 uint8_t driveUntilNewTile(bool* onFirstTile);
-void beginXTicksDrive(uint8_t leadingEncoderLabel, uint8_t max_ticks, bool inReverse);
+void beginXTicksDrive(uint8_t leadingEncoderLabel, int max_ticks, bool inReverse);
 void moveRobotXTicks();
 void driveTestOff();
 void controlLoop(int loopDelayMs);
